+++ CHANGES.rst
@@ -6,6 +6,9 @@ Change Log
 
 Released on XXX, 2013
 
+* Removed ``simpletree`` from the package. The default tree builder is now
+  ``dom`` (using the ``xml.dom.minidom`` implementation).
+
 
 0.95
 ~~~~